Store region inventory data based on universe names instead of world names.
Since regular inventory data is stored based on universe names as well
this should simplify logic.

This would also make it so regions with the same name and universe but in
different worlds use the same inventory, but you're normally not supposed
to be able to create multiple regions with the same name in different
worlds. However this is still possible in some cases.