Implement inventory changing support when moving between inventory regions.
Works when moving in and out of an inventory region and when moving
between inventory regions. Does not support nested inventory regions.

When moving between inventory regions the default inventory for the
world's universe flashes for a brief and often unnoticeable moment,
which may cause problems in rare cases when the player is moving between
regions during a perfectly timed lag-spike.